% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulcanexample.R
\name{vulcanexample}
\alias{vulcanexample}
\title{Function to load an example VULCAN dataset into the workspace}
\usage{
vulcanexample()
}
\value{
A list of components:
\describe{
\item{peakcounts}{A matrix of raw peak counts, peaks as rows, samples as
columns}
\item{peakrpkms}{A matrix of peak RPKMs, peaks as rows, samples as
columns}
\item{samples}{A vector of sample names and conditions}
}
}
\description{
This function loads a dummy ChIP-Seq dataset used for VULCAN examples.
}
\examples{
vobj<-vulcanexample()
names(vobj)
vobj$peakcounts[1:5,]
}
