% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulcansheet.R
\name{vulcansheet}
\alias{vulcansheet}
\title{FUnction to generate an import sheet CSV file for the vulcan test set}
\usage{
vulcansheet(outfile = "sheetfile.csv")
}
\arguments{
\item{outfile}{Path and file name for the CSV sheet file}
}
\value{
nothing
}
\description{
This function generates a user-specific example sheet with appropriate paths
on a dummy ChIP-Seq dataset
}
\examples{
vfile<-"deleteme.csv"
vulcansheet(vfile)
tmp<-read.csv(vfile)
head(tmp)
unlink(vfile)
}
